Rename package for 'roboclaw_node', fixes #7
