+++ setup.cfg
@@ -1,4 +1,4 @@
 [develop]
-script_dir=$base/lib/roboclaw_node_ros
+script_dir=$base/lib/roboclaw_ros
 [install]
-install_scripts=$base/lib/roboclaw_node_ros
+install_scripts=$base/lib/roboclaw_ros
